+++ paper/paper.bib
@@ -1,3 +1,98 @@
+@article{alain2016,
+  author = {Guillaume Alain and Yoshua Bengio},
+  title = {Understanding intermediate layers using linear classifier probes},
+  journal={ArXiv},
+  year={2016},
+  volume={abs/1610.01644}
+}
+
+@incollection{kernelPCA,
+title = {Layer-wise analysis of deep networks with Gaussian kernels},
+author = {Montavon, Gr\'{e}goire and M\"{u}ller, Klaus-Robert and Mikio L. Braun},
+booktitle = {Advances in Neural Information Processing Systems 23},
+editor = {J. D. Lafferty and C. K. I. Williams and J. Shawe-Taylor and R. S. Zemel and A. Culotta},
+pages = {1678--1686},
+year = {2010},
+publisher = {Curran Associates, Inc.},
+url = {http://papers.nips.cc/paper/4061-layer-wise-analysis-of-deep-networks-with-gaussian-kernels.pdf}
+}
+
+@article{gradcamplusplus,
+   title={Grad-CAM++: Generalized Gradient-Based Visual Explanations for Deep Convolutional Networks},
+   ISBN={9781538648865},
+   url={http://dx.doi.org/10.1109/WACV.2018.00097},
+   DOI={10.1109/wacv.2018.00097},
+   journal={2018 IEEE Winter Conference on Applications of Computer Vision (WACV)},
+   publisher={IEEE},
+   author={Chattopadhay, Aditya and Sarkar, Anirban and Howlader, Prantik and Balasubramanian, Vineeth N},
+   year={2018},
+   month={Mar}
+}
+
+
+@article{gradcam,
+   title={Grad-CAM: Visual Explanations from Deep Networks via Gradient-Based Localization},
+   volume={128},
+   ISSN={1573-1405},
+   url={http://dx.doi.org/10.1007/s11263-019-01228-7},
+   DOI={10.1007/s11263-019-01228-7},
+   number={2},
+   journal={International Journal of Computer Vision},
+   publisher={Springer Science and Business Media LLC},
+   author={Selvaraju, Ramprasaath R. and Cogswell, Michael and Das, Abhishek and Vedantam, Ramakrishna and Parikh, Devi and Batra, Dhruv},
+   year={2019},
+   month={Oct},
+   pages={336–359}
+}
+
+
+@misc{svcca2,
+      title={Insights on representational similarity in neural networks with canonical correlation},
+      author={Ari S. Morcos and Maithra Raghu and Samy Bengio},
+      year={2018},
+      eprint={1806.05759},
+      archivePrefix={arXiv},
+      primaryClass={stat.ML}
+}
+
+@article{goingdeeper,
+  author    = {Mats L. Richter and
+               Julius Sch{\"{o}}ning and
+               Ulf Krumnack},
+  title     = {Should You Go Deeper? Optimizing Convolutional Neural Network Architectures
+               without Training by Receptive Field Analysis},
+  journal   = {CoRR},
+  volume    = {abs/2106.12307},
+  year      = {2021},
+  url       = {https://arxiv.org/abs/2106.12307},
+  archivePrefix = {arXiv},
+  eprint    = {2106.12307},
+  timestamp = {Wed, 30 Jun 2021 16:14:10 +0200},
+  biburl    = {https://dblp.org/rec/journals/corr/abs-2106-12307.bib},
+  bibsource = {dblp computer science bibliography, https://dblp.org}
+}
+
+@article{keskar,
+  author    = {Nitish Shirish Keskar and
+               Dheevatsa Mudigere and
+               Jorge Nocedal and
+               Mikhail Smelyanskiy and
+               Ping Tak Peter Tang},
[…]
