+++ paper/paper.bib
@@ -244,4 +244,5 @@ url = {http://papers.neurips.cc/paper/9015-pytorch-an-imperative-style-high-perf
   year={2016},
   volume={},
   number={},
+  doi={10.1109/cvpr.2016.319},
   pages={2921-2929},}
