+++ .github/workflows/tests.yaml
@@ -44,7 +44,7 @@ jobs:
             pip install pandas==1.1.4
             pip install -r requirements/requirements.txt
             pip install -r requirements/docs.txt
-            pip install --upgrade pytest flake8 sphinx
+            pip install --upgrade pytest flake8 sphinx --user
             pip install scipy --force-reinstall # for https://github.com/conda/conda/issues/6396
             pip install .
             conda install pytest
