+++ paper/paper.bib
@@ -169,6 +169,7 @@ author={
   archivePrefix = {arXiv},
   eprint    = {2102.01582},
   timestamp = {Thu, 18 Feb 2021 08:11:35 +0100},
+  doi       = {10.1007/978-3-030-86340-1_11},
   biburl    = {https://dblp.org/rec/journals/corr/abs-2102-01582.bib},
   bibsource = {dblp computer science bibliography, https://dblp.org}
 }
