### 2021-11-06 #3812 @rkurchin
@soroushj and @yzhao062, would you be interested in reviewing this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
