### 2021-11-16 #3812 @rkurchin
@tuchandra and @sambitdash, would you be interested in reviewing this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
