### 2021-12-14 #3992 @whedon
[REVIEW]: Delve: Neural Network Feature Variance Analysis
**Submitting author:** @justinshenk (<a href="http://orcid.org/0000-0002-0664-7337">Justin Shenk</a>)
**Repository:** <a href="https://github.com/delve-team/delve" target ="_blank">https://github.com/delve-team/delve</a>
**Version:** v0.1.49
