### 2021-12-19 bf88887
Update paper.md
A few typographic errors.
