+++ paper/paper.md
@@ -30,33 +30,33 @@ bibliography: paper.bib
 
 # Summary
 Designing neural networks is a complex task.
-Deep neural networks are often refered to as "black box" models - little insight in the function they approximate is gained from looking at the structure of layer outputs.
+Deep neural networks are often referred to as "black box" models - little insight in the function they approximate is gained from looking at the structure of layer outputs.
 ``Delve`` is a tool for looking at how a neural network represents data, and how these representations, or features, change throughout training.
 This tool enables deep learning researchers to understand the limitations and suggest improvements for the design of their networks, such as removing or adding layers.
 
 Several tools exist which allow analyzing neural networks after and during training.
 These techniques can be characterized by their focus on either data or model as well as their level of abstractness.
-Examples for abstract model-oriented techniques are tools for analyzing the sharpness of local optima [@keskar;@sensitivitygoogle], which can be an indicator for the generalizing capeabilities of the trained models.
-In these scenarios the complex of dataset and model is reduced to the error surface, allowing for insights into the differences between different setups.
+Examples for abstract model-oriented techniques are tools for analyzing the sharpness of local optima [@keskar;@sensitivitygoogle], which can be an indicator for the generalizing capabilities of the trained models.
+In these scenarios the complexity of the dataset and model is reduced to the error surface, allowing for insights into the differences between different setups.
 A less abstract data-centric technique GradCam by Selvaraju et al. [@gradcam;@gradcamplusplus], reduces the model to a set of class-activation maps that can be overlayed over individual data points to get an intuitive understanding of the inference process.
-SVCCA [@svcca;@svcca2] can be considered model-centric and a middle ground in terms of abstractness, since it allows the comparative analysis on the features extracted by specific layers.
+SVCCA [@svcca;@svcca2] can be considered model-centric and a middle ground in terms of abstractness, since it allows the comparative analysis of the features extracted by specific layers.
 SVCCA is also relevant from a functional perspective for this work, since it uses singular value decomposition as a core technique to obtain the analysis results.
-Another model-centric tool that allows for a layer-by-layer analysis is logistic regression probes [@alain2016], which utilize logistic regressions trained on the output of a hidden layer to measure the linear seperability of the data and thus the quality of the intermediate solution quality of a classifier model.
+Another model-centric tool that allows for a layer-by-layer analysis is logistic regression probes [@alain2016], which utilize logistic regressions trained on the output of a hidden layer to measure the linear separability of the data and thus the quality of the intermediate solution quality of a classifier model.
 
-The latter is of great importance for this work, since logistic regression probes are often used to compare models and identify the contribution of layers to overall performance [@feature-space;@sizematters;@goingdeeper] and to demonstrate that the saturation metric is capable of showing parameter-inefficienies in neural netwoork architectures.
+The latter is of great importance for this work since logistic regression probes are often used to compare models and identify the contribution of layers to overall performance [@feature-space;@sizematters;@goingdeeper] and to demonstrate that the saturation metric is capable of showing parameter-inefficiencies in neural network architectures.
 
-However, the aforementioned tools have significant limitation in terms of their usefulness in practical application scenarios, where these tools are to be used to improve the performance of a given model.
-In case of data-centric tools like GradCam the solution propagates back to the data, which makes it hard to derive decisions regarding the neural architecture.
-However, the biggest concern in all aforementioned tools are the cost in computational resources and the integration of the analysis into the workflow of a deep learning practicioner.
+However, the aforementioned tools have significant limitations in terms of their usefulness in practical application scenarios, where these tools are to be used to improve the performance of a given model.
[…]
