### 2021-12-19 d5ed8c1
Update paper.md
Fixing a bad grammar checked in.
