### 2021-12-19 #3992 @sambitdash
Hi @rkurchin,

I reviewed the software paper.

There are some typos. They are identified and a PR is submitted for easier recognition and fixing. https://github.com/delve-team/delve/pull/53.

Although the paper talks about the limitations of the other techniques, it does not have a reference to a numerical comparison of those techniques with the proposed technique. If that has not been carried out, maybe that should be mentioned as a scope of further work in the paper.
