### 2021-12-21 #3992 @justinshenk
Hi @sambitdash,
Thank you for your valuable suggestions and improvements. The PR has been merged and a note added in https://github.com/delve-team/delve/commit/811756520cbfd8dce4427c53203ac193f61a94d1 regarding the scope of further work.
