### 2021-12-22 delve-team/delve#54 @dionhaefner
[JOSS review] Pinning Pytorch
Is it really necessary to pin Pytorch to `==1.9.0`? Seems quite restrictive to me, and makes the package harder to install (because if you e.g. do `pip install delve`  and then `pip install torchvision` it gets overwritten again).

(This is a part of the ongoing review at https://github.com/openjournals/joss-reviews/issues/3992)
