### 2021-12-22 delve-team/delve#55 @dionhaefner
[JOSS review] Incorrect qualifiers
Qualifiers in `setup.py`:

```python
        'Programming Language :: Python :: 3.4',
        'Programming Language :: Python :: 3.5',
        'Programming Language :: Python :: 3.6',
```

But since you have `python_requires='>=3.6'`, this should probably be something like `3.6` through `3.10`.

(This is a part of the ongoing review at openjournals/joss-reviews#3992)
