### 2021-12-22 acd2b07
[JOSS review] typos in README
(This is a part of the ongoing review at openjournals/joss-reviews#3992)
