### 2021-12-22 delve-team/delve#57 @dionhaefner
[JOSS review] Test coverage
I suggest adding a service like `codecov` to see how much is actually covered by tests, and adding a badge to the README. There's no shame in not reaching 100% coverage, but if you don't measure it you won't know whether your tests work as intended.

(This is a part of the ongoing review at openjournals/joss-reviews#3992)
