### 2021-12-22 delve-team/delve#58 @dionhaefner
[JOSS review] API
I wonder if `CheckLayerSat` is really the best name for your main tracker object. The imperative sounds more like a function name to me, and `Sat` is so overloaded that it's not obvious what it stands for. I would probably use something like `SaturationTracker` or so.

But I understand that changing names in the public API can be a pain, so if you insist to keep it that's fine with me.

(This is a part of the ongoing review at openjournals/joss-reviews#3992)
