### 2021-12-22 delve-team/delve#59 @dionhaefner
[JOSS review] Doc nitpicks
Things I noticed while reading the docs:

- Spurious `indices and tables` link on the saturation page.
- If I understand correctly then `CheckLayerSat` is the only way your users should interact with the library. In this case there's no need to include anything else in the API reference. Just focus on the essential API and exclude internal objects.
- Broken link on top of Reference page.
- I think the home page of your documentation fills a similar role as the GitHub README, in being the first point of interaction for new users where you should put your best foot forward. Right now your README is a lot more polished, so why not just include the README in the documentation home page and save yourself the hassle of maintaining both separately? (E.g. by converting the README to rst, see [mpi4jax](https://github.com/mpi4jax/mpi4jax) where we use this pattern.)
- I would link more prominently to the integration with the tensorflow playground, which really does a great job of introducing the library! Love the gif.
- Links under "dependencies" are broken (and the whole section is unnecessary IMO).
- Emphasize more clearly what I should read to understand the theory behind Delve. You mention several papers but I think highlighting a specific one could be helpful.

(This is a part of the ongoing review at openjournals/joss-reviews#3992)
