### 2021-12-22 #3992 @dionhaefner
Hey @justinshenk, congratulations on the nice library you and your collaborators have built!

I opened some issues related to the code and docs. On top of that, here are some points I had regarding the paper:

- I think you do a good job of summarizing the status quo, but what I find missing is a clear statement on what delve is trying to be. Is delve a tool to monitor layer saturation? Is it a general monitoring tool for layer statistics? From the REAMDE and docs I thought it was the former, but the paper makes it sounds like the goal is the latter.
- If your goal for delve is to become a general one-stop tool for layer statistics you will need to lay out your plans in a lot more detail. What are the next steps towards this? Adding more layer types can't be the only thing left to do. (I would probably suggest you frame this mostly as a tool for layer saturation though, and the rest as the cherry on top.)
- The section "Eigendecomposition of the feature covariance matrix" does nothing for me personally, because it is too brief to understand what's going on. Either remove it and refer to the paper or at least explain all quantities and operators in the equation.
