### 2021-12-22 delve-team/delve#55 @justinshenk
Thanks! Fixed in https://github.com/delve-team/delve/commit/70b6ddb290622a1938cf4ea3bcf6da6c1b25d268, closing.
