### 2021-12-29 #3992 @MLRichter
Hello,
sorry for the delayed answer during Christmastime.
Hope you all had a great holiday season!
Thank you for your feedback, I have compiled some answers to your remarks:

1. What is delve trying to be?
It is primarily a tool for monitoring saturation and derived information like the intrinsic dimensionality and some visualizations, and this is its primary goal.
It CAN be used for experiment control as well. However, this can be considered as a freebie, that just comes with the logic necessary for the former. There are more advanced solutions for pure experiment control and monitoring like MLFlow and the stuff from weights and biases for example.
This is also the reason why we work with the writer-patterns, which makes it easy to hook delve into any data sink and by extension monitoring system, which we think is important for easy integration into existing code.
That being said, I agree that it should be clarified in the text, that we are mainly providing interesting statistics based on the spectrum of the latent representations.

2. Regarding the topic "Eigendecomposition of the feature covariance matrix"
I agree with the reviewer that this chapter is either undersized or unnecessary.
I think it is essential to elaborate on the way the covariance matrix is computed, since its memory and computationally efficient computation during training is one of the selling points of delve.
Thus, I would like to expand in this chapter and elaborate the computation of the covariance matrix in greater detail. Furthermore highlighting the benefits of our approximation method when it comes to practical usage.
