+++ paper/paper.md
@@ -106,24 +106,26 @@ This would also contradict our goal of seamless integration in existing training
 mini-batches, which are substantially smaller than the entire dataset.
 Therefore, an batch-wise approximation algorithm is used in order to compute the covariance matrix life during training:
 
-
-We compute the covariance matrix $Q(Z_l,Z_l)$, where $Z_l := \sum^{n}_{i=1}(z_{l,i})/n$, using the covariance approximation algorithm between two random variables $X$ and $Y$ with $n$ samples:
+We compute the covariance matrix $Q(Z_l,Z_l)$, where $Z_l$ is the output of a layer $l$ by using the covariance approximation algorithm for two random variables $X$ and $Y$ with $n$ samples:
 $$Q(X, Y) = \frac{\sum^{n}_{i=1} x_i y_i}{n} - \frac{(\sum^{n}_{i=1} x_i)  (\sum^{n}_{i=1} y_i)}{n^2}$$
-We make this computation more efficient by exploiting the shape of the layer output matrix $A_l$:
-We compute $\sum^{n}_{i=1} x_i y_i$ for all feature combinations in layer $l$ by calculating the running squares $\sum^{B}_{b=0}A_{l,b}^T A_{l,b}$ of the batch output matrices $A_{l,b}$ where $b \in \{0,...,B-1\}$ for $B$ batches. We replace $\frac{(\sum^{n}_{i=1} x_i)  (\sum^{n}_{i=1} y_i)}{n^2}$ by the outer product $\bar{A}_l \bigotimes \bar{A}_l$ of the sample mean $\bar{A}_l$.
+The advantage of this method is that only the number of seen samples, the sum of squares and the sum of the variables need to be stored,
+making the memory consumption per layer constant with respect to the size of the dataset.
+By exploiting the shape of the layer output matrix $A_l$ of the layer $l$ we can compute the all covariances simultaniously:
+We can compute $\sum^{n}_{i=1} x_i y_i$ for all feature combinations in layer $l$ by calculating the running squares $\sum^{B}_{b=0}A_{l,b}^T A_{l,b}$ of the batch output matrices $A_{l,b}$ where $b \in \{0,...,B-1\}$ for $B$ batches. We replace $\frac{(\sum^{n}_{i=1} x_i)  (\sum^{n}_{i=1} y_i)}{n^2}$ by the outer product $\bar{A}_l \bigotimes \bar{A}_l$ of the sample mean $\bar{A}_l$.
 This is the running sum of all outputs $z_{l,k}$, where $k \in \{0,...,n\}$ at training time, divided by the total number of training samples $n$.
-The final formula for covariance approximation is then:
+Our formula for a batch-wise approximated covariance matrix can now be written like this:
 $$Q(Z_l, Z_l) = \frac{\sum^{B}_{b=0}A_{l,b}^T A_{l,b}}{n} -(\bar{A}_l \bigotimes \bar{A}_l)$$
-Since we only store the sum of squares, the running mean and the number of observed samples, we require constant memory and computation is done batch-wise.
-The algorithm requires roughly the same number of computations as the processing of a forward pass of the respective layer does; thus we compute saturation after every epoch. The variables are reset at the beginning of each epoch to minimize the bias induced by weight updates during training.
+The batch-wise updating algorithm allows us to integrate the approximation of the covariance matrix as part of the 
+regular forward pass during training and evaluation.
 Our algorithm uses a thread-save common value store on a single compute device or node, which furthermore allows to update the covariance matrix asynchronous when the network is trained in a distributed manner.
+To avoid problems that can be caused rounding errors and numeric instability our implementation of the algorithm exclusivly converts by default all data into 64-bit floating point values.
 
[…]
