### 2022-01-02 #3992 @MLRichter
Happy New Year everyone,
This pull request, contains the reworked text Eigendecomposition:
https://github.com/delve-team/delve/pull/60
Please let me know if this is sufficient!
