### 2022-01-03 #3992 @dionhaefner
Thanks @MLRichter, the section is much better now!

The remaining to-do list from my side is this:

- [ ] Clarify in the paper what Delve is trying to be, i.e.

   > primarily a tool for monitoring saturation and derived information like the intrinsic dimensionality and some visualizations

- [ ] Address the issues on the Delve repository.
- [ ] I noticed quite a few typos in the new section, so please run the manuscript through a spell checker or ask a colleague to provide some copy-editing.

After that we're good to go!
