+++ paper/paper.md
@@ -98,31 +98,25 @@ Several layers are currently supported:
 Additional layers such as PyTorch's ConvTranspose2D are planned for future development (see issue [#43](https://github.com/delve-team/delve/issues/43)).
 
 ## Eigendecomposition of the feature covariance matrix
-The computation of saturation and other related metrics like the intrinsic dimensionality require the covariance matrix
-of the layers output.
-Computing the covariance matrix of a layers output on the training or evaluation set is impractical to do naivly, since
-it would require holding the entire dataset in memory.
-This would also contradict our goal of seamless integration in existing training loops, which commonly operate with
-mini-batches.
-Therefore, a batch-wise approximation algorithm is used in order to compute the covariance matrix life during training:
+The computation of saturation and other related metrics like the intrinsic dimensionality require the covariance matrix of the layers output.
+Computing the covariance matrix of a layer's output on the training or evaluation set is impractical to do naively, since it would require holding the entire dataset in memory.
+This would also contradict our goal of seamless integration in existing training loops, which commonly operate with mini-batches.
+Therefore, a batch-wise approximation algorithm is used to compute the covariance matrix life during training:
 
 We compute the covariance matrix $Q(Z_l,Z_l)$, where $Z_l$ is the output of a layer $l$ by using the covariance approximation algorithm for two random variables $X$ and $Y$ with $n$ samples:
 $$Q(X, Y) = \frac{\sum^{n}_{i=1} x_i y_i}{n} - \frac{(\sum^{n}_{i=1} x_i)  (\sum^{n}_{i=1} y_i)}{n^2}$$
 The advantage of this method is that only the number of seen samples, the sum of squares and the sum of the variables need to be stored,
 making the memory consumption per layer constant with respect to the size of the dataset.
-By exploiting the shape of the layer output matrix $A_l$ of the layer $l$ we can compute all covariances simultaniously:
+By exploiting the shape of the layer output matrix $A_l$ of the layer $l$ we can compute the covariance of all variable pairs simultaneously:
 We can compute $\sum^{n}_{i=1} x_i y_i$ for all feature combinations in layer $l$ by calculating the running squares $\sum^{B}_{b=0}A_{l,b}^T A_{l,b}$ of the batch output matrices $A_{l,b}$ where $b \in \{0,...,B-1\}$ for $B$ batches. We replace $\frac{(\sum^{n}_{i=1} x_i)  (\sum^{n}_{i=1} y_i)}{n^2}$ by the outer product $\bar{A}_l \bigotimes \bar{A}_l$ of the sample mean $\bar{A}_l$.
 This is the running sum of all outputs $z_{l,k}$, where $k \in \{0,...,n\}$ at training time, divided by the total number of training samples $n$.
 Our formula for a batch-wise approximated covariance matrix can now be written like this:
 $$Q(Z_l, Z_l) = \frac{\sum^{B}_{b=0}A_{l,b}^T A_{l,b}}{n} -(\bar{A}_l \bigotimes \bar{A}_l)$$
-The batch-wise updating algorithm allows us to integrate the approximation of the covariance matrix as part of the 
-regular forward pass during training and evaluation.
-Our algorithm uses a thread-save common value store on a single compute device or node, which furthermore allows to update the covariance matrix asynchronous when the network is trained in a distributed manner.
-To avoid problems that can be caused by rounding errors and numeric instability our implementation of the algorithm exclusivly converts by default all data into 64-bit floating point values.
+The batch-wise updating algorithm allows us to integrate the approximation of the covariance matrix as part of the regular forward pass during training and evaluation.
+Our algorithm uses a thread-save common value store on a single compute device or node, which furthermore allows updating the covariance matrix asynchronous when the network is trained in a distributed manner.
[…]
