### 2022-01-05 98b72ba
changed the location of the test folder in the hope of changing the coverage to be correctly measured.
