### 2022-01-05 0c0c9ef
changed how pytest is run to avoid module-confusion
