+++ .github/workflows/tests.yaml
@@ -47,7 +47,7 @@ jobs:
             pip install --upgrade pytest flake8 sphinx --user
             pip install scipy --force-reinstall # for https://github.com/conda/conda/issues/6396
             conda install pytest
-            py.test . --cov-report=xml --cov=delve -vvv
+            python -m pytest tests --cov-report=xml --cov=delve -vvv
       - name: Codecov
         uses: codecov/codecov-action@v2.0.2
         with:
