### 2022-01-05 97d0bbc
added more tests to get more coverage
