### 2022-01-05 81a3e11
added tests for more writers
