### 2022-01-05 08d9c54
added tests for more writers
