### 2022-01-05 delve-team/delve#58 @MLRichter
Hi, I agree, and I think changing it is a good Idea, since the bugfix is trivial.
This is currently implemented in PR as part of a code cleanup effort and waiting for approval.
https://github.com/delve-team/delve/pull/63
