### 2022-01-05 delve-team/delve#54 @MLRichter
Resolved in PR https://github.com/delve-team/delve/pull/62, waiting for approval.
