### 2022-01-05 #3992 @MLRichter
Hi,
just giving you a brief update.
Currently, work is almost finished on these three PRs:
https://github.com/delve-team/delve/pull/61 addressing Issue: https://github.com/delve-team/delve/issues/54
https://github.com/delve-team/delve/pull/63 addressing Issues: https://github.com/delve-team/delve/issues/57 and https://github.com/delve-team/delve/issues/58
https://github.com/delve-team/delve/pull/61 addressing the issues regarding the written paper.

https://github.com/delve-team/delve/issues/59   is the last open issue, will be addressed next?
