### 2022-01-12 #3992 @sambitdash
> There's still [delve-team/delve#59](https://github.com/delve-team/delve/issues/59), but if the authors prefer to address that later it's fine with me.

@dionhaefner, agree with you.
