### 2022-01-12 #3992 @rkurchin
@justinshenk, would you like to address any/all of delve-team/delve#59 prior to concluding this review? A lot of the points there seem like fairly easy fixes...
