+++ docs/source/reference.rst
@@ -1,27 +1,12 @@
 Reference
 =============
 
-PCA Layer Methods
------------------
-
-The following methods are available via :module:`delve.pca_layers`:
-
-.. automodule:: delve.pca_layers
-    :members:
-    :undoc-members:
-    :noindex:
-
 CheckLayerSat
 -------------
 
 ``CheckLayerSat`` provides a hook for PyTorch and extracts metrics during model training.
 
 .. autoclass:: delve.CheckLayerSat
-
-TorchCovarianceMatrix
----------------------
-
-.. autoclass:: delve.torch_utils.TorchCovarianceMatrix
 
 
 API Pages
