+++ docs/source/install.rst
@@ -31,10 +31,10 @@ Installation with pip should also include all dependencies, but a complete list 
 Usage
 -----
 
-Instantiate the :class:`~delve.torchcallback.CheckLayerSat` class where you define your PyTorch training loop, as in the example::
+Instantiate the :class:`~delve.torchcallback.SaturationTracker` class where you define your PyTorch training loop, as in the example::
 
    from torch import nn
-   from delve import CheckLayerSat
+   from delve import SaturationTracker
 
    ...
 
@@ -45,7 +45,7 @@ Instantiate the :class:`~delve.torchcallback.CheckLayerSat` class where you defi
 
 
    layers = [model.conv1, model.linear1]
-   stats = CheckLayerSat('regression/h{}'.format(h),
+   stats = SaturationTracker('regression/h{}'.format(h),
       save_to="plotcsv",
       modules=layers,
       stats=["lsat"]
