+++ .pre-commit-config.yaml
@@ -5,7 +5,6 @@ repos:
     -   id: check-json
     -   id: check-yaml
     -   id: end-of-file-fixer
-    -   id: trailing-whitespace
     -   id: requirements-txt-fixer
 -   repo: https://github.com/pre-commit/mirrors-yapf
     rev: 'v0.31.0'  # Use the sha / tag you want to point at
