### 2022-01-17 delve-team/delve#59 @justinshenk
Thanks for the helpful suggestions! These are addressed in v0.1.49 with commits:
a273dbddce1bc631049ce1ec5b96e5df7329ef66
563c69ea4341884c405753d3b264c6c386109815
6ac1bebf7f49b96688c43742e296402c33cd840a
