+++ README.rst
@@ -1,0 +1,226 @@
+Delve: Deep Live Visualization and Evaluation |logo|
+====================================================
+
+|PyPI version| |Tests| |codecov.io| |License: MIT| |DOI| 
+
+Delve is a Python package for analyzing the inference dynamics of your model.
+
+.. image:: https://raw.githubusercontent.com/justinshenk/playground/master/saturation_demo.gif
+   :alt: playground
+
+Use Delve if you need a lightweight PyTorch extension that: 
+
+- Gives you insight into the inference dynamics of your architecture 
+- Allows you to optimize and adjust neural networks models to your dataset without much trial and error 
+- Allows you to analyze the eigenspaces your data at different stages of inference 
+- Provides you basic tooling for experiment logging
+
+Motivation
+----------
+
+Designing a deep neural network is a trial and error heavy process that
+mostly revolves around comparing performance metrics of different runs.
+One of the key issues with this development process is that the results
+of metrics not really propagate back easily to concrete design
+improvements. Delve provides you with spectral analysis tools that allow
+you to investigate the inference dynamic evolving in the model while
+training. This allows you to spot underutilized and unused layers.
+Mismatches between object size and neural architecture among other
+inefficiencies. These observations can be propagated back directly to
+design changes in the architecture even before the model has fully
+converged, allowing for a quicker and more guided design process.
+
+This work is closely related to Maithra Raghu (Google Brain) et al's work on SVCCA:
+
+- "Maithra Raghu on the differences between wide and deep networks", 2020 `[YouTube] <https://youtu.be/6uPop547u_E?t=970>`_
+- "SVCCA:Singular Vector Canonical Correlation Analysis for Deep Learning and Interpretability", 2017 `[arXiv] <https://arxiv.org/abs/1706.05806>`_
+  
+Installation
+------------
+
+.. code:: bash
+
+   pip install delve
+
+Using Layer Saturation to improve model performance
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+The saturation metric is the core feature of delve. By default
+saturation is a value between 0 and 1.0 computed for any convolutional,
+lstm or dense layer in the network. The saturation describes the
+percentage of eigendirections required for explaining 99% of the
+variance. Simply speaking, it tells you how much your data is “filling
+up” the individual layers inside your model.
+
+In the image below you can see how saturation portraits inefficiencies
+in your neural network. The depicted model is ResNet18 trained on 32
+pixel images, which is way to small for a model with a receptive field
+exceeding 400 pixels in the final layers.
+
+.. image:: https://raw.githubusercontent.com/delve-team/delve/master/images/resnet.PNG
+   :alt: resnet.PNG
+
+To visualize what this poorly chosen input resolution does to the
+inference, we trained logistic regressions on the output of every layer
+to solve the same task as the model. You can clearly see that only the
+first half of the model (at best) is improving the intermedia solutions
+of our logistic regression “probes”. The layers following this are
+contributing nothing to the quality of the prediction! You also see that
+saturation is extremly low for this layers!
+
+We call this a *tail* and it can be removed by either increasing the
+input resolution or (which is more economical) reducing the receptive
+field size to match the object size of your dataset.
+
+.. figure:: https://raw.githubusercontent.com/delve-team/delve/master/images/resnetBetter.PNG
+   :alt: resnetBetter.PNG
+
[…]
