### 2022-01-17 #3992 @justinshenk
Thanks guys for your patience! Corrections made as suggested in the latest version v0.1.49 as described in https://github.com/delve-team/delve/issues/59.
Latest Zenodo DOI for v0.1.49: https://zenodo.org/record/5865465.
