+++ MANIFEST.in
@@ -1,2 +1,3 @@
 include HISTORY.rst
 include LICENSE
+include README.rst
