+++ setup.py
@@ -31,7 +31,7 @@ with open('HISTORY.rst') as history_file:
 requirements = ['tensorboardX', 'tqdm', 'matplotlib', 'pandas', 'torch']
 
 this_dir = path.abspath(path.dirname(__file__))
-with open(os.path.join(this_dir, 'README.md'), encoding='utf-8') as f:
+with open(os.path.join(this_dir, 'README.rst'), encoding='utf-8') as f:
     long_description = f.read()
 
 setup(
