### 2022-01-17 #3992 @rkurchin
A few editorial suggestions/remarks for the manuscript:
* line 44: should read "needs to be able to be used with as little computational and workflow overhead as possible" (first "as" currently missing)
* 48: "known" used twice, remove one of them
* 60: I would rephrase to "`Delve` directly hooks into..."
* 70-71: I would rephrase to "has exploded in recent years"
* 72: Pluralize "Publication"
* 76: "hinting on" should be "hinting at"
* 78: remove "an"
* 82-83: JFYI, it's not necessary to explicitly link the Zenodo here as it will be in the frontmatter of the paper. That being said, it's no problem if you want to!
* 104-105: Phrasing here is a bit confusing because you say you're going to compute Q(Z_l, Z_l) but then show an equation for Q(X,Y). I would rephrase this section to start with that equation and just say something like "We wish to compute the covariance between two random variables X and Y, given by..." and then remark that to avoid having to store all the samples in memory, we approximate by...etc.
* 105-106: I would add somewhere the formal definitions of x_i, y_i for completeness (I recognize it's probably obvious to most readers, but it still feels weird to me not to have it there)
* 119: "asynchronous" should be "asynchronously"
* 120: "numeric" should be "numerical"
* 121: the phrase "by default" should be moved to the end of the sentence (after "floating-point values")
* 126-127: I'm not sure what's happening notationally with x vs. dot, but I interpret both of those as multiplication, so some clarification is needed one why these are different.
