### 2022-01-24 #3992 @MLRichter
https://github.com/delve-team/delve/pull/65

This PR should address all remaining issues.
Please briefly read through my changed made in line 104-105 regarding the covariance matrix and let me know if this satisfies you.
